# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.CreateProjectPayload:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.GetProjectsResponse:
    properties:
      createdAt:
        type: string
      doneTasks:
        type: integer
      id:
        type: string
      inProgressTasks:
        type: integer
      managerId:
        type: string
      name:
        type: string
      status:
        $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_models.Projectstatus'
      teamId:
        type: string
      toDoTasks:
        type: integer
      updatedAt:
        type: string
    type: object
...
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.MessageResponse:
    properties:
      message:
        example: operation completed successfully
        type: string
    type: object
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.ProjectsListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.GetProjectsResponse'
        type: array
      pagination:
        $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.Pagination'
    type: object
...
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.UpdateProjectPayload:
    properties:
      name:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_models.Projectstatus'
        enum:
        - OnHold
        - InProgress
        - Completed
    required:
    - name
    - status
    type: object
...
  github_com_TobiasRV_challenge-fs-senior_internals_models.Projectstatus:
    enum:
    - OnHold
    - InProgress
    - Completed
    type: string
    x-enum-varnames:
    - ProjectstatusOnHold
    - ProjectstatusInProgress
    - ProjectstatusCompleted
...
  github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorBody:
    properties:
      body:
        example: error message
        type: string
    type: object
  github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse:
    properties:
      errors:
        $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorBody'
    type: object
  github_com_TobiasRV_challenge-fs-senior_internals_utils.Pagination:
    properties:
      next_cursor:
        example: eyJpZCI6IjEyMzQ1In0=
        type: string
      prev_cursor:
        example: eyJpZCI6IjEyMzQ1In0=
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for managing teams, projects, tasks, and users
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Challenge FS Senior API
  version: "1.0"
paths:
...
  /projects:
    get:
      consumes:
      - application/json
      description: Get paginated list of projects (Admin or Manager only)
      parameters:
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Filter by team ID
        in: query
        name: teamId
        type: string
      - description: Filter by manager ID
        in: query
        name: managerId
        type: string
      - description: Include task statistics
        in: query
        name: withStats
        type: boolean
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Number of items per page
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.ProjectsListResponse'
        "400":
          description: Bad request - teamId or managerId required
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "403":
          description: Forbidden - Admin or Manager only
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all projects
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: Create a new project (Manager only)
      parameters:
      - description: Project creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.CreateProjectPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.GetProjectsResponse'
        "400":
          description: Validation error or manager not found
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "403":
          description: Forbidden - Manager only
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new project
      tags:
      - Projects
  /projects/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a project by ID (Manager only)
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.MessageResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "403":
          description: Forbidden - Manager only
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a project
      tags:
      - Projects
    put:
      consumes:
      - application/json
      description: Update project information (Manager only, must be the project manager)
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Project update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.UpdateProjectPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.GetProjectsResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "403":
          description: Forbidden - Must be the project manager
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a project
      tags:
      - Projects
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"